added functionality to the login page and made some cosmetic changes

+++ client/src/services/EventsService.js
@@ -1,27 +1,27 @@
-import Api from '@/services/Api'
+import {Api} from '@/services/Api'
 import FormData from 'form-data'
 
 export default {
   fetchEvents () {
-    return Api().get('api/events')
+    return Api.get('api/events')
   },
   addEvent (params) {
-    return Api().post('api/events', params)
+    return Api.post('api/events', params)
   },
   updateEvent (params) {
-    return Api().put(`api/events/${params.id}`, params)
+    return Api.put(`api/events/${params.id}`, params)
   },
   uploadEventImage (params) {
     let form = new FormData()
 
     form.append('eventImg', params.image)
 
-    return Api().patch(`api/events/${params.id}/image`, form)
+    return Api.patch(`api/events/${params.id}/image`, form)
   },
   getEvent (params) {
-    return Api().get(`api/event/${params.id}`)
+    return Api.get(`api/event/${params.id}`)
   },
   deleteEvent (id) {
-    return Api().delete(`api/events/${id}`)
+    return Api.delete(`api/events/${id}`, { headers: { Authorization: localStorage.getItem('token') } })
   }
 }
